Backslash
Indeed the first implementation of quotes wasn't really working properly.
This still reduces the encoding done during a submit but makes sure that
backlashes are handled correctly.
Note however that it'll be good to avoid backslash in file paths as they
cannot be merged with reports from other OS.
Unit test added