Code highlight
Display the original code highlighting only the portion that is not covered.
This works for statements and conditions, not sure on how to display non covered functions.
View still needs works.

This commit changes the way ids are built and fixes #14

It also removes the dependency on clientCode from within the basic interpreter.