Performance improvements
This commit improves performance by allowing to reduce the amount of data
that is included in instrumented files and the amount of data sent to the
node-coverage server with the submit method.

There are two kinds of data:
- "staticInfo": data about the file itself (list of statement ids,
condition ids, function ids, highlighted code)
- "run": info about the code that was actually run. This includes the
counters for each statement, condition or function that were actually
executed at least once.

There is normally no need to include the "staticInfo" part in the instrumented
files and have that info be sent to the browser and back to the server.
It is only needed on the server, along with the "run" part to compute and
display coverage results on the server.

This commit clearly separates those two kinds of information and allows to
create a JSON file containing the static part which will allow not to store
it in instrumented files themselves, thus reducing the size of instrumented
files and improving performance when calling submit.

This commit also removes the stack of all function calls that was built
when function coverage was enabled, as it is usually very big, takes a lot
of time to be sent to the server, and is not very useful. Instead, there
is now a counter for each function in the same way as it is done for
statement coverage.